+++ app/Http/Controllers/RoleController.php
@@ -16,6 +16,8 @@ class RoleController extends Controller
      */
     public function index()
     {
+        $this->authorize('role_access');
+
         $roles = Role::paginate(10);
         return view('roles.index', compact('roles'));
     }
@@ -27,6 +29,8 @@ class RoleController extends Controller
      */
     public function create()
     {
+        $this->authorize('role_create');
+
         $permissions = Permission::all();
         return view('roles.create',compact('permissions'));
     }
@@ -39,6 +43,8 @@ class RoleController extends Controller
      */
     public function store(RoleRequest $request)
     {
+        $this->authorize('role_create');
+
         $role = Role::create(['name' => $request->input('name')]);
         $role->syncPermissions($request->input('permissions'));
         return redirect()->route('roles.index')
@@ -53,6 +59,8 @@ class RoleController extends Controller
      */
     public function show(Role $role)
     {
+        $this->authorize('role_access');
+
         $permissions = $role->permissions;
 
         return view('roles.show', compact('role', 'permissions'));
@@ -66,6 +74,8 @@ class RoleController extends Controller
      */
     public function edit(Role $role)
     {
+        $this->authorize('role_edit');
+
         $permissions = Permission::all();
         return view('roles.edit',compact('role','permissions'));
     }
@@ -79,6 +89,8 @@ class RoleController extends Controller
      */
     public function update(RoleRequest $request, Role $role)
     {
+        $this->authorize('role_edit');
+
         $role->update([
             'name' => $request->name,
         ]);
@@ -96,6 +108,8 @@ class RoleController extends Controller
      */
     public function destroy(Role $role)
     {
+        $this->authorize('role_delete');
+
         $role->delete();
         return response()->json(__('roles.Delete Success'));
     }
